pyzmq-vidwin/pub/requirements.txt
pyzmq
psutil
Pympler
opencv-contrib-python-headless
ffmpeg
docker
pillow==5.4.1
tensorflow==2.1.0
